Backend Python/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pyodbc
openpyxl
python-multipart
reportlab
openpyxl
python-multipart
reportlab